fix: Keep at top of pings (#31)
Fixes #25

If already at the top of the list of pings, when a new ping is added, it keeps being at the top of the list, to ensure the new ping is visible. But it won't show it if you were not at the top (this is usually the expected behaviour).