fix: Allow flaky UI tests to pass in 3 attempts (#67)
I haven't been able to make those tests fail after dozens of times. But I've known snackbars and activity lifecycles to be somewhat flaky things to assert.

Our helper testing library has this annotation that attempts a test multiple times, to see if it eventually passes. I know it's not an ideal solution, but it's a way for us to keep those particular checks. Otherwise we'll probably need to re-write the test to assert state in some other way that's not through UI.

Closes #61